Update 102-building-a-basic-workflow.md
This might not be the best way to do it, but the instructions advise to drag the new transition to the Remove Group Member action, but that action isn't created until the next step. I moved it under point 10 in the next step, but I'm not sure that's the best place for it.